Applied State Hoisting to SearchAppBar.kt

+++ app/src/main/java/com/example/composereciepeeapp/presentation/ui/recipe_list/RecipeListViewModel.kt
@@ -57,7 +57,7 @@ class RecipeListViewModel @Inject constructor(
         onQueryChanged(category)
     }
 
-    fun onChangedCategoryScrollPosition(position: Int){
+    fun onChangedCategoryScrollPosition(position: Float){
         categoryScrollPosition = position.toFloat()
     }
 }
